requirements.txt
airbyte==0.31.3
FastAPI==0.115.9
uvicorn==0.34.3
PyJWT==2.8.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
vanna==0.7.9
chromadb
google-genai
google-cloud-aiplatform==1.96.0
onnxruntime==1.22.0
langchain-google-genai<1.0.0
langchain

pytest==7.4.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
httpx
requests-mock==1.11.0

torch>=2.2.0,<2.8.0
transformers>=4.35.0,<5.0.0
tokenizers>=0.15.0,<1.0.0
scikit-learn>=1.5.2,<2.0.0
numpy>=1.21.0,<1.27.0

matplotlib==3.9.2 
pandas==2.2.3